Make command ring synchronous, use correct CLOSE code
We don't care about asynchronous execution on the command ring - we do
it so rarely anyway; this makes sure that consumers don't accidentally
get ahead of the ball (eg. by issuing TD doorbells before the TD ring is
up).

This also uses the correct code for closing a TD ring, so now closing
and reopening RPC or AT rings works correctly.