Improve upload interaction with the tty layer
The TTY layer expects the write_room() value to be accurate, but given
we output in discrete pages, it's impossible to be right. This patch
adds a byte ring buffer so that write() has a linear space to go into.
This also results in write consolidation when the TD ring buffer is
full.

When a line discipline is engaged (ie. PPP), write_room() is never
called. Instead, when the buffer fills up, the ldisc will retry a
few times before giving up, potentially for a long period. This results
in terrible upload performance. The ldisc provides a write_wakeup()
method to signal when it can resume, so we use this when necessary (ie.
when a write failed because the buffer was full).