Added net_device_stats to push IO stats on the device
The xmm7360 module works great. But it seems dead on the UI
because there are not stats available on the device. The system
monitor etc can see the device but can't show the current statistics
like speed, packets etc.

This patch fixes that and addds net_device_stats to the driver
and keeps track of the packets and number of bytes Tx/Rx'ed.

Signed-off-by: Himanshu Chauhan <hschauhan@nulltrace.org>